Changing package name to org.artrev.compareverifier
